Update callback function parameter types to use `uint16_t` instead of `uint8_t`.

+++ examples/List/List.ino
@@ -22,8 +22,8 @@
 #define CLEAR 46     // NUMPAD .
 
 // Declare the calbacks
-void colorsCallback(uint8_t pos);
-void numsCallback(uint8_t pos);
+void colorsCallback(uint16_t pos);
+void numsCallback(uint16_t pos);
 
 // Declare the array
 extern String colors[];
@@ -74,12 +74,12 @@ void loop() {
 }
 
 // Define the calbacks
-void colorsCallback(uint8_t pos) {
+void colorsCallback(uint16_t pos) {
     // do something with the index
     Serial.println(colors[pos]);
 }
 
-void numsCallback(uint8_t pos) {
+void numsCallback(uint16_t pos) {
     // do something with the index
     Serial.println(nums[pos]);
 }
